admin-backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyotp>=2.9.0
qrcode[pil]>=7.4.2
redis>=5.0.1
httpx>=0.26.0
python-multipart>=0.0.6
email-validator>=2.1.0
tenacity>=8.2.0

# TON/USDT Deposit System
aiogram>=3.0.0
pytoniq>=0.1.0
Pillow>=10.0.0
